services/dast-engine/requirements.txt
fastapi
uvicorn
docker
pydantic
python-owasp-zap-v2.4
requests